% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerWrapper.R
\name{Power}
\alias{Power}
\title{Power}
\usage{
Power(
  data = NULL,
  version = "0.97.1",
  alpha = 0.05,
  alternative = "twoSided",
  baselineProportion = 0.5,
  calculation = "sampleSize",
  comparisonProportion = 0.6,
  effectDirection = "greater",
  effectDirectionSyntheticDataset = "less",
  effectSize = 0.5,
  firstGroupMean = 0,
  firstGroupSd = 1,
  logSampleSize = TRUE,
  plotHeight = 320,
  plotWidth = 480,
  populationSd = 1,
  power = 0.9,
  powerByEffectSize = TRUE,
  powerBySampleSize = FALSE,
  powerContour = TRUE,
  powerDemonstration = FALSE,
  sampleSize = 20,
  sampleSizeRatio = 1,
  saveDataset = FALSE,
  savePath = "",
  secondGroupMean = 0,
  secondGroupSd = 1,
  seed = 1,
  setSeed = FALSE,
  test = "independentSamplesTTest",
  testValue = 0,
  text = TRUE,
  varianceRatio = 2
)
}
\description{
Power
}
